requirements.txt
asgiref==3.7.2
Authlib==1.2.1
cffi==1.15.1
cryptography==41.0.3
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.3
django-static-md5url==0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
gunicorn==21.2.0
packaging==23.1
psycopg==3.1.9
pycparser==2.21
PyJWT==1.7.1
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
